ocr/requirements.txt
openai==1.12.0
boto3==1.34.51
Pillow==10.2.0